Restructure tests in preparation to add model tests
