--- pipeline/scraper/psql/__init__.py
@@ -1,9 +1,0 @@
-"""
-Import functions into a package namespace
-"""
-
-from .psql import connect_to_db
-from .psql import insert_to_app
-from .psql import insert_to_dependencies
-from .psql import insert_to_package
-from .psql import update_package_metadata
